# backendRL/src/main/java/org/example/backendrl/config/api-docs.yaml
openapi: 3.1.0
info:
  title: RecordingLogs API
  description: "Music rating and recommendation platform: genres, artists, albums,\
    \ songs, ratings, reviews, album lists, favorites and follows."
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/reviews/{reviewId}:
    get:
      tags:
      - Reviews
      summary: Get a review by id
      operationId: getReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Review found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumReviewResponse"
    put:
      tags:
      - Reviews
      summary: Edit a review
      operationId: updateReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumReviewRequest"
        required: true
      responses:
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: User is not the author of the review
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Review updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumReviewResponse"
    delete:
      tags:
      - Reviews
      summary: Delete a review
      operationId: deleteReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Review not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Review deleted
...
  /api/albums/{id}:
    get:
      tags:
      - Albums
      summary: Get an album by id
      operationId: getAlbum
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Album found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
    put:
      tags:
      - Albums
      summary: Update an album
      operationId: updateAlbum
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumRequest"
        required: true
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Album updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
    delete:
      tags:
      - Albums
      summary: Delete an album
      operationId: deleteAlbum
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Album deleted
...
  /api/lists/{listId}/albums/{albumId}:
    post:
      tags:
      - Lists
      summary: Add an album to a list
      operationId: addAlbum
      parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumListItemRequest"
        required: true
      responses:
        "404":
          description: List or album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: User is not the owner of the list
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Album is already in the list
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: "Album added, returns the updated list"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumListResponse"
    delete:
      tags:
      - Lists
      summary: Remove an album from a list
      operationId: removeAlbum
      parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: List not found or album not in list
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Album removed
...
  /api/albums:
    get:
      tags:
      - Albums
      summary: Get all albums
      operationId: getAllAlbums
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumResponse"
    post:
      tags:
      - Albums
      summary: Create an album
      operationId: createAlbum
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Album created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
...
  /api/albums/{albumId}/reviews:
    get:
      tags:
      - Reviews
      summary: Get all reviews of an album
      operationId: getReviewsForAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of reviews
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumReviewResponse"
    post:
      tags:
      - Reviews
      summary: Write a review for an album
      operationId: createReview
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumReviewRequest"
        required: true
      responses:
        "404":
          description: Album or user not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: User already reviewed this album
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Review created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumReviewResponse"
...
  /api/users/{userId}/reviews:
    get:
      tags:
      - Reviews
      summary: Get all reviews written by a user
      operationId: getReviewsForUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of reviews
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumReviewResponse"
...
  /api/songs/{songId}/albums:
    get:
      tags:
      - Songs
      summary: Get all albums containing a song
      operationId: getAlbumsForSong
      parameters:
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummaryResponse"
  /api/reviews:
    get:
      tags:
      - Reviews
      summary: "Get all reviews, newest first"
      operationId: getAllReviews
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of reviews
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumReviewResponse"
...
  /api/genres/{genreId}/albums:
    get:
      tags:
      - Genres
      summary: Get all albums of a genre
      operationId: getAlbumsForGenre
      parameters:
      - name: genreId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Genre not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummaryResponse"
...
  /api/artists/{artistId}/albums:
    get:
      tags:
      - Artists
      summary: Get all albums of an artist
      operationId: getAlbumsForArtist
      parameters:
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Artist not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummaryResponse"
...
components:
  schemas:
...
    AlbumReviewRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 200
          minLength: 0
        content:
          type: string
          maxLength: 10000
          minLength: 0
      required:
      - content
      - title
      - userId
    AlbumReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        albumId:
          type: integer
          format: int64
        albumTitle:
          type: string
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AlbumListItemResponse:
      type: object
      properties:
        albumId:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
        position:
          type: integer
          format: int32
    AlbumListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        albums:
          type: array
          items:
            $ref: "#/components/schemas/AlbumListItemResponse"
...
    AlbumRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
      required:
      - title
    AlbumResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
...
    AlbumListItemRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        position:
          type: integer
          format: int32
          minimum: 1
      required:
      - userId
...
    AlbumSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
...